backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
pdfplumber==0.10.3
pdf2image==1.16.3
Pillow==10.1.0
PyPDF2==3.0.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
requests==2.31.0
python-multipart==0.0.6
werkzeug==3.0.1
